Drag-off, water loss, plotting
Fixed a couple bugs and added a few more; currently, the reverse
mode doesn't reproduce the forward mode results consistently,
which is likely buried somewhere in the escape rate calculation.
This commit allows for water to be lost both by H abstraction
and by drag-off, as well as O evolution in response to water
loss.