Consistent escape regime logic fixes
Some of the edge cases were behaving strangely, which required
digging in to the escipe regime/limit calculations slightly. The
code now runs more smoothly through most of the given scenarios.
As an additional improvement, the code now finds the maximum water
fraction (from Noack et al., 2016) for the stated planet mass and
radius, and sets the envelope fraction according to that amount.
This allows for a more consistent parameter sweep (caveat emptor).