Exploratory efforts to quantify Kzz
I coded up a version of Ackerman and Marley (2001)'s Kzz, based
on the description in Gao et al. (2018); the problem is that I
don't have a good sense of the convective heat flux, and I worry
that the radiative flux from sigma*T**4 isn't sufficiently precise
to get even an order-of-magnitude Kzz term. but, it's at least 1E6
cm2/s?